Script will work for seeing systemInfo
